application/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.3.4
python-dateutil==2.9.0.post0
fhir.resources==7.1.0
requests==2.31.0
